Create Accordian Script(Show hide)
With this code, we use a loop to add click event listeners to each heading element and its corresponding hidden text within the ACF repeater. Clicking on a heading will toggle the visibility of the associated hidden text.

Make sure to replace 'repeater_field_name' and 'subfield_name' with the actual names you've set for your ACF fields in the WordPress admin. Additionally, ensure that you've properly enqueued your JavaScript file in your WordPress theme for it to take effect.